[Issue-166] Make RequestBuilder abstract
RequestBuilder can be overrided by sub builders of http methods.
RequestBuilder should be abstract, and its properties are accessable from the sub builders.
Also, the subs should support the method-chain style.
Generic <B> is added for which subs keeps the style.
And it is a type of a sub builder which overrides RequestBuilder.