[issue-87] Fix an issue that there is possibility that duplicate filenames are generated
This issue have never happened yet, but it is possible to occur if AUIL
and Volley are used simultaneously.
If Volley and AUIL are using some disc cache and share the same cache
directory each other, then all of two libraries will put or get a same
cache filename and actual file at the same url.
Because the format of AUIL cache is different from volley's, this
situation can cause an error when one of them reads shared(same) cache file.
To avoid this sharing problem, it must follow the below solutions.
(1) the scope UniversalDiscCache should not be public. Because the
delegate of UniversalDiscCache can be shared with AUIL.
(2) All of disc caches in this library should customized file name
generator other than AUIL's generator.