[issue-193] Fix crash/OOM in DiskBasedCache
This is a copy of an original commit of aosp volley.
The original is in the link,
https://android.googlesource.com/platform/frameworks/volley/+/5bd5325bd61414480091b262204f1a45d57c1884%5E%21/#F0
,
and the commit message is as below.

Fix crash/OOM in DiskBasedCache

DiskBasedCache.writeHeader() can throw an IOException and swallow it.
This causes the
cache entry to be partially written followed by the entry's data. When
reading back,
it is possible to read back the data segment (garbled part) when
attempting to parse
the cache header.